processor-sgx/requirements.txt
kafka-python
elasticsearch>=6.0.0,<7.0.0
